Food Vendors now persist! Read this for debugging:
To load the vendor data, browse to localhost:8888/sneaky_xpress/updateData in your web browser. This will start the remote job on the (debug) server so you don't have to wait for a month for it to update!

Known quirk: The application is consistently reporting that ~97 vendors are removed and 109 vendors are added. I don't know why fewer vendors are being removed than added - keep an eye out for what's causing this!