Sorry, had to change the structure of the Content interface.
The main difference is that the Content clasess are responsible for presenting the data, and the *ServiceImpl classes are only resposible for providing server data. Please look at the BrowseVendors* for an example (it's a really simple one).

I will try not to change it any more, but this time it needed to be done (since we cannot use UI elements in Servelets).